+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### Next (10)
-- 
+- Add user score filter to media lists
 
 ### 0.44 (9)
 - Add role in media to character details